feat: mangarr (#36)
* initial version of mangarr

* workflow updates

* hotfix: disabled specific tests

* hotfix: disabled tests altogether

* hotfix: disable publish to services

* fix: brought back publish to custom endpoint

* chore: version

* feat: add workflow dispatch option

* fix: vulnerable dependency

* feat: add BencodeNET

* refactor: parse torrent with BencodeNET and base64

* chore: disable issue templates for now

* Update FUNDING.yml

* feat: cover image in archive (#20)

* feat: comic info (#21)

* Add comic info data structure

* Add creating and adding comic info

* Extract download functionality to catch errors

* feat: temporarily remove volume number (#22)

* fix: small fixes (#23)

* Formatting

* Fixed web entry in ComicInfo

* Deleted most indexers (#25)

* Deleted most indexers

* Deleted only Newznab file from Newzban folder

* Changed Torznab display name to Sourcerarr

* Remove usenet indexer section from frontend

* feat: removed SceneMapping functionality (#26)

* Replaced all functionality with stubs

* Removed unused classes

* Removed scheduled task

* Removed Xem

* Removed unused file

* feat: properly render manga description (#27)

* Added interweave package

* Use Markup to render overview

* Render in add new series modal

* Render for episode summary if any

* Render in series details

* Removed some logs

* fix: tests (#28)

* Delete xem tests

* Delete scene mapping tests

* Delete indexer tests

* Update integration tests

* fix: decode entities in title (#29)

* Decode name when mapping series

* More decoding of name

* fix: option for using alternate titles for searching (#30)

* Decode name when mapping series

* More decoding of name

* Add UseAlternateTitlesForSearch property to Series

* Only add alternate titles if requested

* Add extra migration step to add to db

* Update frontend to have access to new value/toggle

* Fixes for backend

* feat: remove download clients (#32)

* Remove all download clients

* Update references

* Remove usenet section from frontend

* fix: remove download client tests (#34)

* Remove download client tests

* Fix migration test

* Update workflows

+++ src/NzbDrone.Core.Test/ImportListTests/ImportListSyncServiceFixture.cs
@@ -77,7 +77,7 @@ namespace NzbDrone.Core.Test.ImportListTests
 
             Mocker.GetMock<ISeriesService>()
                   .Setup(v => v.AllSeriesTvdbIds())
-                  .Returns(new List<int>());
+                  .Returns(new List<long>());
 
             Mocker.GetMock<ISeriesService>()
                 .Setup(v => v.GetAllSeries())
@@ -137,7 +137,7 @@ namespace NzbDrone.Core.Test.ImportListTests
         {
             Mocker.GetMock<ISeriesService>()
                   .Setup(v => v.AllSeriesTvdbIds())
-                  .Returns(new List<int> { _list1Series.First().TvdbId });
+                  .Returns(new List<long> { _list1Series.First().TvdbId });
         }
 
         private void WithExcludedSeries()
