+++ src/NzbDrone.Core/Exceptions/SeriesNotFoundException.cs
@@ -4,21 +4,21 @@ namespace NzbDrone.Core.Exceptions
 {
     public class SeriesNotFoundException : NzbDroneException
     {
-        public int TvdbSeriesId { get; set; }
+        public long TvdbSeriesId { get; set; }
 
-        public SeriesNotFoundException(int tvdbSeriesId)
+        public SeriesNotFoundException(long tvdbSeriesId)
             : base(string.Format("Series with tvdbid {0} was not found, it may have been removed from TheTVDB.", tvdbSeriesId))
         {
             TvdbSeriesId = tvdbSeriesId;
         }
 
-        public SeriesNotFoundException(int tvdbSeriesId, string message, params object[] args)
+        public SeriesNotFoundException(long tvdbSeriesId, string message, params object[] args)
             : base(message, args)
         {
             TvdbSeriesId = tvdbSeriesId;
         }
 
-        public SeriesNotFoundException(int tvdbSeriesId, string message)
+        public SeriesNotFoundException(long tvdbSeriesId, string message)
             : base(message)
         {
             TvdbSeriesId = tvdbSeriesId;
