+++ src/NzbDrone.Core/ImportLists/Exclusions/ImportListExclusion.cs
@@ -4,7 +4,7 @@ namespace NzbDrone.Core.ImportLists.Exclusions
 {
     public class ImportListExclusion : ModelBase
     {
-        public int TvdbId { get; set; }
+        public long TvdbId { get; set; }
         public string Title { get; set; }
     }
 }
